pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coin-trading"
version = "0.1.0"
description = "Reproducible RL trading framework scaffold"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "pydantic>=2.6,<3",
  "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
